requirements.txt
streamlit
pandas
pillow
scikit-learn
matplotlib
seaborn
streamlit-elements
